+++ wms-web-springboot/src/main/java/com/teeny/wms/web/model/request/OutputPickingRequestEntity.java
@@ -16,7 +16,7 @@ public class OutputPickingRequestEntity {
 
     private int id;
     private int detailId;
-    private int number;
+    private float number;
     private List<OutputPickingEntity> list;
 
     public int getId() {
@@ -35,11 +35,11 @@ public class OutputPickingRequestEntity {
         this.detailId = detailId;
     }
 
-    public int getNumber() {
+    public float getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(float number) {
         this.number = number;
     }
